pyproject.toml
[project]
name = "mercury-cli"
version = "1.3.3"
description = "A CLI wrapper for Mercury-OCIP"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Jordan Prescott", email = "devteam@fourteenip.com" },
    { name = "Michael Clarke", email = "devteam@fourteenip.com" },
    { name = "Billy Kelly", email = "devteam@fourteenip.com" },
]
dependencies = [
    "prompt-toolkit>=3.0.52",
    "yaspin>=3.3.0",
    "prompt-toolkit-action-completer @ git+https://github.com/Fourteen-IP/prompt-toolkit-action-completer.git@1.2.3",
    "argparse>=1.4.0",
    "asciimatics>=1.15.0",
    "mercury-ocip @ git+https://github.com/Fourteen-IP/mercury-ocip.git@v0.1.2-beta",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.16",
    "mkdocs-minify-html-plugin>=0.3.5",
    "mkdocstrings[python]>=0.30.0",
    "pymdown-extensions>=10.16.1",
    "mkdocs-literate-nav>=0.6.2",
    "mkdocs-asciinema-player>=0.18.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
pythonpath = ["src"]
addopts = "-ra -q"
filterwarnings = ["ignore::DeprecationWarning", "ignore::UserWarning"]

[tool.ruff]
exclude = ["tests/*", "test_*.py"]

[project.scripts]
mercury-cli = "mercury_cli.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["mercury_cli*"]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
branch = "main"
commit_message = "chore(release): {version}"
commit_parser = "conventional"
tag_format = "v{version}"
upload_to_pypi = false
upload_to_vcs_release = true

[tool.semantic_release.changelog]
exclude_commit_patterns = ["^chore", "^docs", "^style", "^test"]

[tool.semantic_release.commit_parser_options]
allowed_tags = [
    "feat",
    "fix",
    "perf",
    "docs",
    "style",
    "refactor",
    "test",
    "build",
    "ci",
    "chore",
]
minor_tags = ["feat"]
patch_tags = ["fix", "perf", "refactor"]